refactor various notification modules
