+++ src/app/feature/chat/services/online/online-chat.service.ts
@@ -1,10 +1,11 @@
 import { Injectable } from "@angular/core";
-import { BehaviorSubject, Observable, from, Subscription } from "rxjs";
+import { BehaviorSubject, Observable, from as observableFrom, Subscription } from "rxjs";
 import { ChatMessage, IChatService } from "../../models";
 import { Capacitor } from "@capacitor/core";
-import { NotificationService } from "src/app/shared/services/notification/notification.service";
+import { PushNotificationService } from "src/app/shared/services/notification/push-notification.service";
 import { takeWhile } from "rxjs/operators";
 import { convertFromRapidProMsg } from "../../utils/message.converter";
+import { RapidproNotificationService } from "src/app/shared/services/notification/rapidpro-notification.service";
 
 @Injectable({
   providedIn: "root",
@@ -16,7 +17,10 @@ export class OnlineChatService implements IChatService {
 
   private ready$ = new BehaviorSubject<boolean>(false);
 
-  constructor(private notificationService: NotificationService) {
+  constructor(
+    private rapidproNotificationService: RapidproNotificationService,
+    private pushNotificationService: PushNotificationService
+  ) {
     this.init();
   }
 
@@ -27,7 +31,7 @@ export class OnlineChatService implements IChatService {
 
   private async init() {
     if (Capacitor.isNative) {
-      await this.notificationService.init();
+      await this.rapidproNotificationService.init();
       this.ready$.next(true);
     }
   }
@@ -38,7 +42,9 @@ export class OnlineChatService implements IChatService {
     if (this.subscription) {
       this.subscription.unsubscribe();
     }
-    this.subscription = this.notificationService.messages$.subscribe(
+    // TODO - filter out only rapidpro push notifications (in case other notifications added since first implementation)
+    // TODO - refactor to only rely on rapidpro notifications service (not generic push)
+    this.subscription = this.pushNotificationService.messages$.subscribe(
       (rpMsg) => {
         convertFromRapidProMsg(rpMsg).then((chatMessage) => {
           this.messages$.next(this.messages$.getValue().concat([chatMessage]));
@@ -50,11 +56,11 @@ export class OnlineChatService implements IChatService {
     );
     // TODO - can we send generic start message or do we need to manually track trigger phrases
     // (as had been done previously)
-    await this.notificationService.sendRapidproMessage(flowName);
+    await this.rapidproNotificationService.sendRapidproMessage(flowName);
     return this.messages$;
   }
 
   public sendMessage(message: ChatMessage): Observable<any> {
-    return from(this.notificationService.sendRapidproMessage(message.text));
+    return observableFrom(this.rapidproNotificationService.sendRapidproMessage(message.text));
   }
 }
