+++ src/app/feature/notification/notifications-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from "@angular/core";
 import { Routes, RouterModule } from "@angular/router";
 
-import { CampaignDebugPage } from "./pages/campaign-debug/campaign-debug.page";
+import { NotificationsDebugPage } from "../notifications/pages/notifications-debug/notifications-debug.page";
 
 const routes: Routes = [
   {
     path: "",
-    component: CampaignDebugPage,
+    component: NotificationsDebugPage,
   },
 ];
 
@@ -14,4 +14,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
 })
-export class CampaignPageRoutingModule {}
+export class NotificationsRoutingModule {}
