--- src/app/shared/services/notification/notification.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { NotificationService } from './notification.service';
-
-describe('NotificationService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: NotificationService = TestBed.get(NotificationService);
-    expect(service).toBeTruthy();
-  });
-});
